refactor(routes): extract app info route
